--- src/lib/findMe.js
@@ -1,13 +1,0 @@
-export function findMe(callbackSuccess, callbackError = null) {
-  if ('geolocation' in navigator) {
-    const geoOptions = {
-      enableHighAccuracy: true,
-      maximumAge: 30000,
-      timeout: 27000,
-    }
-
-    navigator.geolocation.getCurrentPosition(callbackSuccess, console.log, geoOptions)
-  } else if (callbackError) {
-    callbackError('Permisition denied')
-  }
-}
