+++ src/styles/FormInput.module.css
@@ -17,8 +17,8 @@ input {
 .attachButton {
   height: 32px;
   width: 40px;
-  margin: 5px 15px;
-  background: url(https://image.flaticon.com/icons/svg/116/116312.svg);
+  margin: 10px 1px;
+  background: url(https://image.flaticon.com/icons/svg/130/130884.svg);
   background-repeat: no-repeat;
 }
 
@@ -105,14 +105,14 @@ input {
 }
 
 .dropOut {
-  height: 0;
+  height: 50px;
+  width: 0;
   transition-duration: 0.3s;
-  border-radius: 20px 20px 0 0;
-  background-color: #29384b;
-  position: absolute;
-  bottom: 50px;
-  left: 10px;
-  z-index: 1;
+  background-color: rgba(7, 27, 175, 0.2);
+  position: relative;
+  bottom: 0px;
+  left: 0px;
+  z-index: 0;
   overflow: hidden;
 }
 
@@ -120,7 +120,7 @@ input {
   width: 100%;
   height: 100%;
   display: flex;
-  flex-flow: column;
+  flex-flow: row;
   align-items: center;
 }
